+++ spotifywebapipython/const.py
@@ -4,7 +4,7 @@
 # constants are placed in this file if they are used across multiple files.
 # the only exception to this is for the VERSION constant, which is placed here for convenience.
 
-VERSION:str = "1.0.256"
+VERSION:str = "1.0.257"
 """ 
 Current version of the Spotify Client Python3 Library. 
 """
@@ -50,6 +50,11 @@ Default country code to use if a country code was not supplied on a request and 
 UserProfile.Country value is not set (e.g. public access token in effect. 
 
 Default value is "US".
+"""
+
+SPOTIFY_DESKTOP_APP_CLIENT_DISPLAY_NAME:str = 'Spotify Desktop App Client (%s)'
+"""
+Spotify Desktop Application client display name (e.g. `Spotify Desktop App Client (%s)`).
 """
 
 SPOTIFY_DESKTOP_APP_CLIENT_ID:str = '65b708073fc0480ea92a077233ca87bd'